# docs/technical/03_api_specification.generated.yaml
components:
  schemas:
...
    Body_upload_generic_file_api_v1_media_file_post:
      properties:
        file:
          contentMediaType: application/octet-stream
          title: File
          type: string
      required:
      - file
      title: Body_upload_generic_file_api_v1_media_file_post
      type: object
    Body_upload_orphan_photo_api_v1_media_orphans__orphan_id__photo_post:
      properties:
        file:
          contentMediaType: application/octet-stream
          title: File
          type: string
      required:
      - file
      title: Body_upload_orphan_photo_api_v1_media_orphans__orphan_id__photo_post
      type: object
...
    FileUploadResponse:
      description: 'Thin metadata blob the frontend hands straight to

        POST /orphans/{id}/documents (or any other attach endpoint).'
      properties:
        file_mime_type:
          title: File Mime Type
          type: string
        file_name:
          title: File Name
          type: string
        file_size_bytes:
          title: File Size Bytes
          type: integer
        file_url:
          title: File Url
          type: string
      required:
      - file_url
      - file_name
      - file_size_bytes
      - file_mime_type
      title: FileUploadResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    MarketingChannelCreate:
      properties:
        channel_type:
          anyOf:
          - enum:
            - digital_marketing
            - committee
            - website
            - branch
            - social_media
            - partnership
            - other
            type: string
          - type: 'null'
          title: Channel Type
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        name_ar:
          maxLength: 255
          minLength: 1
          title: Name Ar
          type: string
        name_en:
          anyOf:
          - maxLength: 255
            type: string
          - type: 'null'
          title: Name En
      required:
      - name_ar
      title: MarketingChannelCreate
      type: object
    MarketingChannelRead:
      properties:
        channel_type:
          anyOf:
          - enum:
            - digital_marketing
            - committee
            - website
            - branch
            - social_media
            - partnership
            - other
            type: string
          - type: 'null'
          title: Channel Type
        created_at:
          format: date-time
          title: Created At
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        id:
          format: uuid
          title: Id
          type: string
        name_ar:
          maxLength: 255
          minLength: 1
          title: Name Ar
          type: string
        name_en:
          anyOf:
          - maxLength: 255
            type: string
          - type: 'null'
          title: Name En
        status:
          enum:
          - active
          - suspended
          - archived
          title: Status
          type: string
      required:
      - name_ar
      - id
      - status
      - created_at
      title: MarketingChannelRead
      type: object
    MarketingChannelUpdate:
      properties:
        channel_type:
          anyOf:
          - enum:
            - digital_marketing
            - committee
            - website
            - branch
            - social_media
            - partnership
            - other
            type: string
          - type: 'null'
          title: Channel Type
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        name_ar:
          anyOf:
          - maxLength: 255
            minLength: 1
            type: string
          - type: 'null'
          title: Name Ar
        name_en:
          anyOf:
          - maxLength: 255
            type: string
          - type: 'null'
          title: Name En
        status:
          anyOf:
          - enum:
            - active
            - suspended
            - archived
            type: string
          - type: 'null'
          title: Status
      title: MarketingChannelUpdate
      type: object
    MediaUploadResponse:
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        file_size_bytes:
          title: File Size Bytes
          type: integer
        file_url:
          title: File Url
          type: string
        id:
          format: uuid
          title: Id
          type: string
        media_type:
          const: photo
          title: Media Type
          type: string
        moderation_status:
          title: Moderation Status
          type: string
        orphan_id:
          format: uuid
          title: Orphan Id
          type: string
      required:
      - id
      - orphan_id
      - media_type
      - file_url
      - file_size_bytes
      - moderation_status
      - created_at
      title: MediaUploadResponse
      type: object
...
    OrphanPhoto:
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        file_size_bytes:
          title: File Size Bytes
          type: integer
        file_url:
          title: File Url
          type: string
        id:
          format: uuid
          title: Id
          type: string
        moderation_status:
          title: Moderation Status
          type: string
        presigned_url:
          title: Presigned Url
          type: string
      required:
      - id
      - file_url
      - presigned_url
      - file_size_bytes
      - moderation_status
      - created_at
      title: OrphanPhoto
      type: object
...
    Page_MarketingChannelRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MarketingChannelRead'
          title: Items
          type: array
        limit:
          title: Limit
          type: integer
        offset:
          title: Offset
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - total
      - limit
      - offset
      title: Page[MarketingChannelRead]
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
      type: oauth2
info:
  title: Rufaqaa API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/marketing-channels:
    get:
      operationId: list_marketing_channels_api_v1_marketing_channels_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      - in: query
        name: include_inactive
        required: false
        schema:
          default: false
          title: Include Inactive
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_MarketingChannelRead_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: List Marketing Channels
      tags:
      - marketing-channels
    post:
      operationId: create_marketing_channel_api_v1_marketing_channels_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarketingChannelCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketingChannelRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Create Marketing Channel
      tags:
      - marketing-channels
  /api/v1/marketing-channels/{channel_id}:
    delete:
      operationId: archive_marketing_channel_api_v1_marketing_channels__channel_id__delete
      parameters:
      - in: path
        name: channel_id
        required: true
        schema:
          format: uuid
          title: Channel Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Archive Marketing Channel
      tags:
      - marketing-channels
    get:
      operationId: get_marketing_channel_api_v1_marketing_channels__channel_id__get
      parameters:
      - in: path
        name: channel_id
        required: true
        schema:
          format: uuid
          title: Channel Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketingChannelRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Get Marketing Channel
      tags:
      - marketing-channels
    patch:
      operationId: update_marketing_channel_api_v1_marketing_channels__channel_id__patch
      parameters:
      - in: path
        name: channel_id
        required: true
        schema:
          format: uuid
          title: Channel Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarketingChannelUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketingChannelRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Update Marketing Channel
      tags:
      - marketing-channels
...
  /api/v1/media/file:
    post:
      description: 'Stage a file in object storage and return its s3:// URL plus

        metadata. The frontend two-step upload-then-attach flow calls this

        first, then passes the returned fields to an endpoint that records

        the attachment (e.g. POST /orphans/{id}/documents).'
      operationId: upload_generic_file_api_v1_media_file_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_generic_file_api_v1_media_file_post'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Upload Generic File
      tags:
      - media
  /api/v1/media/orphans/{orphan_id}/photo:
    post:
      operationId: upload_orphan_photo_api_v1_media_orphans__orphan_id__photo_post
      parameters:
      - in: path
        name: orphan_id
        required: true
        schema:
          format: uuid
          title: Orphan Id
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_orphan_photo_api_v1_media_orphans__orphan_id__photo_post'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaUploadResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Upload Orphan Photo
      tags:
      - media
  /api/v1/media/orphans/{orphan_id}/photos:
    get:
      description: 'Photos attached to an orphan, newest first. Each row carries a

        fresh presigned URL so the UI can render the image without exposing

        the bucket.'
      operationId: list_orphan_photos_api_v1_media_orphans__orphan_id__photos_get
      parameters:
      - in: path
        name: orphan_id
        required: true
        schema:
          format: uuid
          title: Orphan Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/OrphanPhoto'
                title: Response List Orphan Photos Api V1 Media Orphans  Orphan Id  Photos Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: List Orphan Photos
      tags:
      - media
  /api/v1/media/{media_id}/url:
    get:
      description: Return a short-lived presigned URL for an uploaded media object.
      operationId: get_media_presigned_url_api_v1_media__media_id__url_get
      parameters:
      - in: path
        name: media_id
        required: true
        schema:
          format: uuid
          title: Media Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                title: Response Get Media Presigned Url Api V1 Media  Media Id  Url Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Get Media Presigned Url
      tags:
      - media
...